minimal_test.py
#!/usr/bin/env python3
"""
Minimal Enterprise Visual DNA System Test
========================================

Simplified version to test core functionality without potential blocking
"""

import sys
import os

def test_minimal_system():
    """Test minimal system functionality"""
    print("🔧 MINIMAL ENTERPRISE TEST")
    print("=" * 40)
    
    # Test 1: Basic imports
    print("\n📦 Testing imports...")
    
    try:
        import numpy as np
        print("✅ NumPy available")
    except:
        print("❌ NumPy missing")
    
    try:
        from PIL import Image
        print("✅ PIL available")
    except:
        print("❌ PIL missing")
    
    # Test 2: VDN Format
    print("\n🗜️ Testing VDN...")
    try:
        sys.path.insert(0, ".")
        import vdn_format
        vdn = vdn_format.VDNFormat()
        test_data = b"test"
        compressed = vdn.compress(test_data)
        print(f"✅ VDN: {len(test_data)} -> {len(compressed)} bytes")
    except Exception as e:
        print(f"❌ VDN error: {e}")
    
    # Test 3: Twmrto
    print("\n🧠 Testing Twmrto...")
    try:
        import twmrto_compression
        comp = twmrto_compression.TwmrtoCompressor()
        result = comp.compress_text("test")
        print(f"✅ Twmrto: 'test' -> '{result}'")
    except Exception as e:
        print(f"❌ Twmrto error: {e}")
    
    # Test 4: Simple visualization
    print("\n🎨 Testing visualization...")
    try:
        import codebase_relationship_analyzer as cra
        
        # Quick file count
        python_files = [f for f in os.listdir(".") if f.endswith(".py")]
        print(f"✅ Found {len(python_files)} Python files")
        
        # Try simple analysis
        if len(python_files) > 0:
            print("✅ Ready for visualization")
        
    except Exception as e:
        print(f"❌ Visualization error: {e}")
    
    print("\n🎯 Minimal test complete")
    return True

def create_next_iteration_plan():
    """Create plan for next iteration"""
    print("\n📋 NEXT ITERATION PLAN")
    print("-" * 30)
    print("1. Fix enterprise system threading issues")
    print("2. Implement non-blocking initialization")
    print("3. Add graceful degradation modes")
    print("4. Create production-safe error handling")
    print("5. Optimize memory usage and performance")
    print("6. Build web interface for visualization")
    print("7. Generate interstellar communication packages")
    print("8. Deploy to production environment")

if __name__ == "__main__":
    try:
        test_minimal_system()
        create_next_iteration_plan()
        print("\n✅ System functional - ready for next iteration!")
        
    except KeyboardInterrupt:
        print("\n⏹️ Test interrupted by user")
    except Exception as e:
        print(f"\n❌ Test failed: {e}")
        import traceback
        traceback.print_exc()
